backend-django/AHP/requirements.txt
asgiref==3.7.2
astroid==3.0.3
black==24.1.1
click==8.1.7
colorama==0.4.6
dill==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
psycopg2==2.9.9
pylint==3.0.3
pytz==2024.1
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2024.1